--- src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer.Common/Resources/xlf/SR.ja.xlf
@@ -1,22 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<xliff xmlns="urn:oasis:names:tc:xliff:document:1.2" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" version="1.2" xsi:schemaLocation="urn:oasis:names:tc:xliff:document:1.2 xliff-core-1.2-transitional.xsd">
-  <file datatype="xml" source-language="en" target-language="ja" original="../SR.resx">
-    <body>
-      <trans-unit id="File_Externally_Modified">
-        <source>File was externally modified: {0}</source>
-        <target state="translated">ファイルが外部で変更されました: {0}</target>
-        <note />
-      </trans-unit>
-      <trans-unit id="Language_Services_Missing_Service">
-        <source>Razor language services not configured properly, missing language service '{0}'.</source>
-        <target state="translated">Razor 言語サービスが正しく構成されていません。言語サービス '{0}' がありません。</target>
-        <note />
-      </trans-unit>
-      <trans-unit id="Unknown_ProjectChangeKind">
-        <source>Unknown ProjectChangeKind {0}</source>
-        <target state="translated">不明な ProjectChangeKind {0}</target>
-        <note />
-      </trans-unit>
-    </body>
-  </file>
-</xliff>
