--- src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer.Common/Resources/xlf/SR.zh-Hant.xlf
@@ -1,22 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<xliff xmlns="urn:oasis:names:tc:xliff:document:1.2" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" version="1.2" xsi:schemaLocation="urn:oasis:names:tc:xliff:document:1.2 xliff-core-1.2-transitional.xsd">
-  <file datatype="xml" source-language="en" target-language="zh-Hant" original="../SR.resx">
-    <body>
-      <trans-unit id="File_Externally_Modified">
-        <source>File was externally modified: {0}</source>
-        <target state="translated">已在外部修改檔案: {0}</target>
-        <note />
-      </trans-unit>
-      <trans-unit id="Language_Services_Missing_Service">
-        <source>Razor language services not configured properly, missing language service '{0}'.</source>
-        <target state="translated">未正確設定 Razor 語言服務，遺漏語言服務 '{0}'。</target>
-        <note />
-      </trans-unit>
-      <trans-unit id="Unknown_ProjectChangeKind">
-        <source>Unknown ProjectChangeKind {0}</source>
-        <target state="translated">未知的 ProjectChangeKind {0}</target>
-        <note />
-      </trans-unit>
-    </body>
-  </file>
-</xliff>
