+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Common/AdhocLanguageServices.cs
