+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Common/AdhocServices.cs
