+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Common/AdhocWorkspaceServices.cs
