+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Common/BackgroundDocumentGenerator.cs
