+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Common/DefaultProjectEngineFactory.cs
