+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Common/DefaultRemoteTextLoaderFactory.cs
