+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Common/HostDocumentComparer.cs
