+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Common/RazorFileChangeKind.cs
