+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Common/RemoteTextLoaderFactory.cs
