+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/DocumentProcessedListener.cs
